semantic_serach_streamlit/requirements.txt
streamlit
openai
sentence-transformers
pinecone-client
BeautifulSoup4